Get metadata objects from a WCSClient object.

Get metadata objects from a WCSClient object. <WCSCoverageSummary> in particular can be used to extract further metadata about individual coverages.

Usage

emdn_get_coverage_summaries(wcs, coverage_ids)

emdn_get_coverage_summaries_all(wcs)

emdn_get_coverage_ids(wcs)

emdn_has_extent_type(
  wcs,
  coverage_ids,
  type = c("temporal", "vertical", "geographic")
)

Arguments

wcs

A WCSClient R6 object, created with function emdn_init_wcs_client.

coverage_ids

character vector of coverage IDs.

type

a coverage dimension type. One of "temporal", "vertical" or "geographic".

Value

emdn_get_coverage_summaries: returns a list of objects of class <WCSCoverageSummary> for each coverage_id provided.

emdn_get_coverage_summaries_all: returns a list of objects of class <WCSCoverageSummary> for each coverage avalable through the service.

emdn_get_coverage_ids returns a character vector of coverage ids.

Functions

emdn_get_coverage_summaries: Get summaries for specific coverages.

emdn_get_coverage_summaries_all: Get summaries for all available coverages from a service.

emdn_get_coverage_ids: Get coverage IDs for all available coverages from a service.

emdn_has_extent_type: check whether a coverage has a particular dimension.

Examples

wcs <- emdn_init_wcs_client(service = "biology")
#> ✔ WCS client created succesfully
#> ℹ Service: <https://geo.vliz.be/geoserver/Emodnetbio/wcs>
#> ℹ Service: "2.0.1"
emdn_get_coverage_summaries_all(wcs)
#> [[1]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__ratio_large_to_small_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[2]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__aca_spp_19582016_L1
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[3]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__cal_fin_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[4]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__cal_hel_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[5]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__met_luc_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[6]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__oit_spp_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[7]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__tem_lon_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[8]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__chli_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[9]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__tot_lar_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[10]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__tot_sma_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
cov_ids <- emdn_get_coverage_ids(wcs)
emdn_get_coverage_summaries(wcs, cov_ids[1:2])
#> [[1]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__ratio_large_to_small_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[2]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__aca_spp_19582016_L1
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: NULL
#>     dimensions: NULL
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#>