Get EMODnet WCS service and available coverage information.

Get EMODnet WCS service and available coverage information.

Usage

emdn_get_wcs_info(
  wcs = NULL,
  service = NULL,
  service_version = c("2.0.1", "2.1.0", "2.0.0", "1.1.1", "1.1.0"),
  logger = c("NONE", "INFO", "DEBUG")
)

emdn_get_all_wcs_info(logger = c("NONE", "INFO", "DEBUG"))

emdn_get_coverage_info(
  wcs = NULL,
  service = NULL,
  coverage_ids,
  service_version = c("2.0.1", "2.1.0", "2.0.0", "1.1.1", "1.1.0"),
  logger = c("NONE", "INFO", "DEBUG")
)

Arguments

wcs

A WCSClient R6 object, created with function emdn_init_wcs_client.

service

the EMODnet OGC WCS service name. For available services, see emdn_wcs().

service_version

the WCS service version. Defaults to "2.0.1".

logger

character string. Level of logger: 'NONE' for no logger, 'INFO' to get ows4R logs, 'DEBUG' for all internal logs (such as as Curl details)

coverage_ids

character vector of coverage IDs.

Value

emdn_get_wcs_info & emdn_get_wcs_info return a list of service level metadata, including a tibble containing coverage level metadata for each coverage available from the service. emdn_get_coverage_info returns a list containing a tibble of more detailed metadata for each coverage specified.

emdn_get_wcs_info / emdn_get_all_wcs_info

emdn_get_wcs_info and emdn_get_all_wcs_info return a list with the following metadata:

data_source: the EMODnet source of data.

service_name: the EMODnet WCS service name.

service_url: the EMODnet WCS service URL.

service_title: the EMODnet WCS service title.

service_abstract: the EMODnet WCS service abstract.

service_access_constraits: any access constraints associated with the EMODnet WCS service.

service_fees: any access fees associated with the EMODnet WCS service.

service_type: the EMODnet WCS service type.

coverage_details: a tibble of details of each coverage available through EMODnet WCS service:

coverage_id: the coverage ID.

dim_n: the number of coverage dimensions

dim_names: the coverage dimension names, units (in brackets) and types.

extent: the coverage extent (xmin, ymin, xmax and ymax).

crs: the coverage CRS (Coordinate Reference System).

wgs84_bbox: the coverage extent (xmin, ymin, xmax and ymax) in WGS84 (EPSG:4326) CRS coordinates.

temporal_extent: the coverage temporal extent (min - max), NA if coverage contains no temporal dimension.

vertical_extent: the coverage vertical extent (min - max), NA if coverage contains no vertical dimension.

subtype: the coverage subtype.

emdn_get_coverage_info

emdn_get_coverage_info returns a tibble with a row for each coverage specified and columns with the following details:

data_source: the EMODnet source of data.

service_name: the EMODnet WCS service name.

service_url: the EMODnet WCS service URL.

coverage_ids: the coverage ID.

band_description: the description of the data contained in the band of the coverage.

band_uom: the unit of measurement of the data contained in the band of the coverage.

constraint: the range of values of the data contained in the band of the coverage.

nil_value: the nil value of the data contained in the band of the coverage.

grid_size: the spatial size of the coverage grid (ncol x nrow).

resolution: the spatial resolution (pixel size) of the coverage grid in the CRS units of measurement (size in the x dimension x size in the y dimension).

dim_n: the number of coverage dimensions

dim_names: the coverage dimension names, units (in brackets) and types.

extent: the coverage extent (xmin, ymin, xmax and ymax).

crs: the coverage CRS (Coordinate Reference System).

wgs84_bbox: the coverage extent (xmin, ymin, xmax and ymax) in WGS84 (EPSG:4326) CRS coordinates.

temporal_extent: the coverage temporal extent (min - max), NA if coverage contains no temporal dimension.

vertical_extent: the coverage vertical extent (min - max), NA if coverage contains no vertical dimension.

subtype: the coverage subtype.

fn_seq_rule: the function describing the sequence rule which specifies the relationship between the axes of data and coordinate system axes.

fn_start_point: the location of the origin of the data in the coordinate system.

For additional details on WCS metadata, see the GDAL wiki section on WCS Basics and GDAL

Details

To minimize the number of requests sent to webservices, these functions use memoise to cache results inside the active R session. To clear the cache, re-start R or run memoise::forget(emdn_get_wcs_info)/memoise::forget(emdn_get_coverage_info)

Functions

emdn_get_wcs_info: Get info on all coverages from am EMODnet WCS service.

emdn_get_all_wcs_info: Get metadata on all services and all available coverages from each service.

emdn_get_coverage_info: Get metadata for specific coverages. Requires a WCSClient R6 object as input.

Examples

# Get information from a wcs object.
wcs <- emdn_init_wcs_client(service = "seabed_habitats")
#> ✔ WCS client created succesfully
#> ℹ Service: <https://ows.emodnet-seabedhabitats.eu/geoserver/emodnet_open_maplibrary/wcs>
#> ℹ Service: "2.0.1"
emdn_get_wcs_info(wcs)
#> $data_source
#> [1] "emodnet_wcs"
#> 
#> $service_name
#> [1] "seabed_habitats"
#> 
#> $service_url
#> [1] "https://ows.emodnet-seabedhabitats.eu/geoserver/emodnet_open_maplibrary/wcs"
#> 
#> $service_title
#> [1] "EMODnet Seabed Habitats Map Library WCS Service"
#> 
#> $service_abstract
#> [1] "WCS end-point for individual habitat maps and models held by EMODnet Seabed Habitats.\r\n\r\nIncludes all data with access limitations of \"View and Download\" (excluding \"View only\" data)"
#> 
#> $service_access_constraits
#> [1] "None"
#> 
#> $service_fees
#> [1] "None"
#> 
#> $service_type
#> [1] "urn:ogc:service:wcs"
#> 
#> $coverage_details
#> # A tibble: 3 × 9
#>   coverage_id         dim_n dim_n…¹ extent crs   wgs84…² tempo…³ verti…⁴ subtype
#>   <chr>               <int> <chr>   <chr>  <chr> <chr>   <chr>   <chr>   <chr>  
#> 1 emodnet_open_mapli…     2 x(m):g… 34317… EPSG… -6.56,… NA      NA      Rectif…
#> 2 emodnet_open_mapli…     2 x(m):g… 34317… EPSG… -6.56,… NA      NA      Rectif…
#> 3 emodnet_open_mapli…     2 x(m):g… 34314… EPSG… -6.56,… NA      NA      Rectif…
#> # … with abbreviated variable names ¹​dim_names, ²​wgs84_bbox, ³​temporal_extent,
#> #   ⁴​vertical_extent
#> 
# Get information using a service name.
emdn_get_wcs_info(service = "biology")
#> ✔ WCS client created succesfully
#> ℹ Service: <https://geo.vliz.be/geoserver/Emodnetbio/wcs>
#> ℹ Service: "2.0.1"
#> $data_source
#> [1] "emodnet_wcs"
#> 
#> $service_name
#> [1] "biology"
#> 
#> $service_url
#> [1] "https://geo.vliz.be/geoserver/Emodnetbio/wcs"
#> 
#> $service_title
#> [1] "EMODnet Biology"
#> 
#> $service_abstract
#> [1] "The EMODnet Biology products include a set of gridded map layers showing the average abundance of marine species for different time windows (seasonal, annual) using geospatial modelling. The spatial modelling tool used to calculate the gridded abundance maps is based on DIVA. DIVA (Data-Interpolating Variational Analysis) is a tool to create gridded data sets from discrete point measurements of the ocean. For the representation of time dynamics, it was decided to produce gridded maps for sliding time windows, e.g. combining one or more years  in one gridded map, so that relatively smooth animated GIF presentations can be produced that show the essential change over time. EMODnet Biology’s data products include the Operational Ocean Products and Services (OOPS), harvested by ICES."
#> 
#> $service_access_constraits
#> [1] "NONE"
#> 
#> $service_fees
#> [1] "NONE"
#> 
#> $service_type
#> [1] "urn:ogc:service:wcs"
#> 
#> $coverage_details
#> # A tibble: 10 × 9
#>    coverage_id        dim_n dim_n…¹ extent crs   wgs84…² tempo…³ verti…⁴ subtype
#>    <chr>              <int> <chr>   <chr>  <chr> <chr>   <chr>   <chr>   <chr>  
#>  1 Emodnetbio__ratio…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#>  2 Emodnetbio__aca_s…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#>  3 Emodnetbio__cal_f…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#>  4 Emodnetbio__cal_h…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#>  5 Emodnetbio__met_l…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#>  6 Emodnetbio__oit_s…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#>  7 Emodnetbio__tem_l…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#>  8 Emodnetbio__chli_…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#>  9 Emodnetbio__tot_l…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#> 10 Emodnetbio__tot_s…     3 lat(de… -75.0… EPSG… -75.05… 1958-0… NA      Rectif…
#> # … with abbreviated variable names ¹​dim_names, ²​wgs84_bbox, ³​temporal_extent,
#> #   ⁴​vertical_extent
#> 
# Get detailed info for specific coverages from wcs object
coverage_ids <- c(
  "emodnet_open_maplibrary__mediseh_cora",
  "emodnet_open_maplibrary__mediseh_posidonia"
)
emdn_get_coverage_info(wcs = wcs,
                       coverage_ids = coverage_ids)
#> # A tibble: 2 × 20
#>   data_…¹ servi…² servi…³ cover…⁴ band_…⁵ band_…⁶ const…⁷ nil_va…⁸ dim_n dim_n…⁹
#>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>      <dbl> <int> <chr>  
#> 1 emodne… https:… seabed… emodne… GRAY_I… W.m-2.… 0, 1    -3.40e38     2 x(m):g…
#> 2 emodne… https:… seabed… emodne… Probab… W.m-2.… -3.402… -1.00e 4     2 x(m):g…
#> # … with 10 more variables: grid_size <chr>, resolution <chr>, extent <chr>,
#> #   crs <chr>, wgs84_extent <chr>, temporal_extent <chr>,
#> #   vertical_extent <chr>, subtype <chr>, fn_seq_rule <chr>,
#> #   fn_start_point <chr>, and abbreviated variable names ¹​data_source,
#> #   ²​service_name, ³​service_url, ⁴​coverage_id, ⁵​band_description, ⁶​band_uom,
#> #   ⁷​constraint, ⁸​nil_value, ⁹​dim_names