Get metadata objects from a WCSClient object.

Get metadata objects from a WCSClient object. <WCSCoverageSummary> in particular can be used to extract further metadata about individual coverages.

Usage

emdn_get_coverage_summaries(wcs, coverage_ids)

emdn_get_coverage_summaries_all(wcs)

emdn_get_coverage_ids(wcs)

emdn_has_dimension(wcs, coverage_ids, type = c("temporal", "vertical"))

emdn_get_coverage_dim_coefs(
  wcs,
  coverage_ids,
  type = c("temporal", "vertical")
)

Arguments

wcs

A WCSClient R6 object, created with function emdn_init_wcs_client.

coverage_ids

character vector of coverage IDs.

type

a coverage dimension type. One of "temporal" or "vertical".

Value

emdn_get_coverage_summaries: returns a list of objects of class <WCSCoverageSummary> for each coverage_id provided.

emdn_get_coverage_summaries_all: returns a list of objects of class <WCSCoverageSummary> for each coverage avalable through the service.

emdn_get_coverage_ids returns a character vector of coverage ids.

emdn_get_coverage_dim_coefs returns a list containing a vector of coefficients for each coverage requested.

a list containing a vector of coefficients for each coverage requested.

Functions

emdn_get_coverage_summaries: Get summaries for specific coverages.

emdn_get_coverage_summaries_all: Get summaries for all available coverages from a service.

emdn_get_coverage_ids: Get coverage IDs for all available coverages from a service.

emdn_has_dimension: check whether a coverage has a temporal or vertical dimension.

emdn_get_coverage_dim_coefs: Get temporal or vertical coefficients for a coverage.

Examples

wcs <- emdn_init_wcs_client(service = "biology")
#> ✔ WCS client created succesfully
#> ℹ Service: <https://geo.vliz.be/geoserver/Emodnetbio/wcs>
#> ℹ Service: "2.0.1"
cov_ids <- emdn_get_coverage_ids(wcs)
cov_ids
#>  [1] "Emodnetbio__ratio_large_to_small_19582016_L1_err"
#>  [2] "Emodnetbio__aca_spp_19582016_L1"                 
#>  [3] "Emodnetbio__cal_fin_19582016_L1_err"             
#>  [4] "Emodnetbio__cal_hel_19582016_L1_err"             
#>  [5] "Emodnetbio__met_luc_19582016_L1_err"             
#>  [6] "Emodnetbio__oit_spp_19582016_L1_err"             
#>  [7] "Emodnetbio__tem_lon_19582016_L1_err"             
#>  [8] "Emodnetbio__chli_19582016_L1_err"                
#>  [9] "Emodnetbio__tot_lar_19582016_L1_err"             
#> [10] "Emodnetbio__tot_sma_19582016_L1_err"             
emdn_has_dimension(wcs,
                     cov_ids,
                     type = "temporal")
#> Emodnetbio__ratio_large_to_small_19582016_L1_err 
#>                                             TRUE 
#>                  Emodnetbio__aca_spp_19582016_L1 
#>                                             TRUE 
#>              Emodnetbio__cal_fin_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__cal_hel_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__met_luc_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__oit_spp_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__tem_lon_19582016_L1_err 
#>                                             TRUE 
#>                 Emodnetbio__chli_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__tot_lar_19582016_L1_err 
#>                                             TRUE 
#>              Emodnetbio__tot_sma_19582016_L1_err 
#>                                             TRUE 
emdn_has_dimension(wcs,
                     cov_ids,
                     type = "vertical")
#> Emodnetbio__ratio_large_to_small_19582016_L1_err 
#>                                            FALSE 
#>                  Emodnetbio__aca_spp_19582016_L1 
#>                                            FALSE 
#>              Emodnetbio__cal_fin_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__cal_hel_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__met_luc_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__oit_spp_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__tem_lon_19582016_L1_err 
#>                                            FALSE 
#>                 Emodnetbio__chli_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__tot_lar_19582016_L1_err 
#>                                            FALSE 
#>              Emodnetbio__tot_sma_19582016_L1_err 
#>                                            FALSE 
emdn_get_coverage_summaries(wcs, cov_ids[1:2])
#> [[1]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__ratio_large_to_small_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[2]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__aca_spp_19582016_L1
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
emdn_get_coverage_summaries_all(wcs)
#> [[1]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__ratio_large_to_small_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[2]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__aca_spp_19582016_L1
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[3]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__cal_fin_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[4]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__cal_hel_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[5]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__met_luc_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[6]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__oit_spp_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[7]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__tem_lon_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[8]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__chli_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[9]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__tot_lar_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
#> [[10]]
#> <WCSCoverageSummary>
#>   Inherits from: <OGCAbstractObject>
#>   Public:
#>     BoundingBox: list
#>     CoverageId: Emodnetbio__tot_sma_19582016_L1_err
#>     CoverageSubtype: RectifiedGridCoverage
#>     CoverageSubtypeParent: NULL
#>     ERROR: function (text) 
#>     INFO: function (text) 
#>     WARN: function (text) 
#>     WGS84BoundingBox: list
#>     attrs: list
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     element: AbstractObject
#>     encode: function (addNS = TRUE, geometa_validate = TRUE, geometa_inspire = FALSE, 
#>     getBoundingBox: function () 
#>     getClass: function () 
#>     getClassName: function () 
#>     getCoverage: function (bbox = NULL, crs = NULL, time = NULL, elevation = NULL, 
#>     getCoverageStack: function (time = NULL, elevation = NULL, bbox = NULL, filename_handler = NULL, 
#>     getDescription: function () 
#>     getDimensions: function () 
#>     getId: function () 
#>     getNamespaceDefinition: function (recursive = FALSE) 
#>     getSubtype: function () 
#>     getSubtypeParent: function () 
#>     getWGS84BoundingBox: function () 
#>     initialize: function (xmlObj, capabilities, serviceVersion, owsVersion, logger = NULL) 
#>     isFieldInheritedFrom: function (field) 
#>     logger: function (type, text) 
#>     loggerType: NULL
#>     namespace: OWSNamespace, R6
#>     verbose.debug: FALSE
#>     verbose.info: FALSE
#>     wrap: FALSE
#>   Private:
#>     capabilities: WCSCapabilities, OWSCapabilities, OGCAbstractObject, R6
#>     description: WCSCoverageDescription, GMLCOVAbstractCoverage, GMLAbstractCoverage, GMLAbstractFeature, GMLAbstractObject, ISOAbstractObject, geometaLogger, R6
#>     dimensions: list
#>     fetchCoverageSummary: function (xmlObj, serviceVersion, owsVersion) 
#>     fromComplexTypes: function (value) 
#>     owsVersion: 2.0
#>     system_fields: verbose.info verbose.debug loggerType wrap element names ...
#>     url: https://geo.vliz.be/geoserver/Emodnetbio/wcs
#>     version: 2.0.1
#>     xmlElement: AbstractObject
#>     xmlExtraNamespaces: NULL
#>     xmlNamespacePrefix: OWS
#>     xmlNodeToCharacter: function (x, ..., indent = "", tagSeparator = "\n") 
#> 
emdn_get_coverage_dim_coefs(wcs,
                            cov_ids[1:2],
                            type = "temporal")
#> $Emodnetbio__ratio_large_to_small_19582016_L1_err
#>   [1] "1958-02-16T00:00:00.000Z" "1958-05-16T00:00:00.000Z"
#>   [3] "1958-08-16T00:00:00.000Z" "1958-11-16T00:00:00.000Z"
#>   [5] "1959-02-16T00:00:00.000Z" "1959-05-16T00:00:00.000Z"
#>   [7] "1959-08-16T00:00:00.000Z" "1959-11-16T00:00:00.000Z"
#>   [9] "1960-02-16T00:00:00.000Z" "1960-05-16T00:00:00.000Z"
#>  [11] "1960-08-16T00:00:00.000Z" "1960-11-16T00:00:00.000Z"
#>  [13] "1961-02-16T00:00:00.000Z" "1961-05-16T00:00:00.000Z"
#>  [15] "1961-08-16T00:00:00.000Z" "1961-11-16T00:00:00.000Z"
#>  [17] "1962-02-16T00:00:00.000Z" "1962-05-16T00:00:00.000Z"
#>  [19] "1962-08-16T00:00:00.000Z" "1962-11-16T00:00:00.000Z"
#>  [21] "1963-02-16T00:00:00.000Z" "1963-05-16T00:00:00.000Z"
#>  [23] "1963-08-16T00:00:00.000Z" "1963-11-16T00:00:00.000Z"
#>  [25] "1964-02-16T00:00:00.000Z" "1964-05-16T00:00:00.000Z"
#>  [27] "1964-08-16T00:00:00.000Z" "1964-11-16T00:00:00.000Z"
#>  [29] "1965-02-16T00:00:00.000Z" "1965-05-16T00:00:00.000Z"
#>  [31] "1965-08-16T00:00:00.000Z" "1965-11-16T00:00:00.000Z"
#>  [33] "1966-02-16T00:00:00.000Z" "1966-05-16T00:00:00.000Z"
#>  [35] "1966-08-16T00:00:00.000Z" "1966-11-16T00:00:00.000Z"
#>  [37] "1967-02-16T00:00:00.000Z" "1967-05-16T00:00:00.000Z"
#>  [39] "1967-08-16T00:00:00.000Z" "1967-11-16T00:00:00.000Z"
#>  [41] "1968-02-16T00:00:00.000Z" "1968-05-16T00:00:00.000Z"
#>  [43] "1968-08-16T00:00:00.000Z" "1968-11-16T00:00:00.000Z"
#>  [45] "1969-02-16T00:00:00.000Z" "1969-05-16T00:00:00.000Z"
#>  [47] "1969-08-16T00:00:00.000Z" "1969-11-16T00:00:00.000Z"
#>  [49] "1970-02-16T00:00:00.000Z" "1970-05-16T00:00:00.000Z"
#>  [51] "1970-08-16T00:00:00.000Z" "1970-11-16T00:00:00.000Z"
#>  [53] "1971-02-16T00:00:00.000Z" "1971-05-16T00:00:00.000Z"
#>  [55] "1971-08-16T00:00:00.000Z" "1971-11-16T00:00:00.000Z"
#>  [57] "1972-02-16T00:00:00.000Z" "1972-05-16T00:00:00.000Z"
#>  [59] "1972-08-16T00:00:00.000Z" "1972-11-16T00:00:00.000Z"
#>  [61] "1973-02-16T00:00:00.000Z" "1973-05-16T00:00:00.000Z"
#>  [63] "1973-08-16T00:00:00.000Z" "1973-11-16T00:00:00.000Z"
#>  [65] "1974-02-16T00:00:00.000Z" "1974-05-16T00:00:00.000Z"
#>  [67] "1974-08-16T00:00:00.000Z" "1974-11-16T00:00:00.000Z"
#>  [69] "1975-02-16T00:00:00.000Z" "1975-05-16T00:00:00.000Z"
#>  [71] "1975-08-16T00:00:00.000Z" "1975-11-16T00:00:00.000Z"
#>  [73] "1976-02-16T00:00:00.000Z" "1976-05-16T00:00:00.000Z"
#>  [75] "1976-08-16T00:00:00.000Z" "1976-11-16T00:00:00.000Z"
#>  [77] "1977-02-16T00:00:00.000Z" "1977-05-16T00:00:00.000Z"
#>  [79] "1977-08-16T00:00:00.000Z" "1977-11-16T00:00:00.000Z"
#>  [81] "1978-02-16T00:00:00.000Z" "1978-05-16T00:00:00.000Z"
#>  [83] "1978-08-16T00:00:00.000Z" "1978-11-16T00:00:00.000Z"
#>  [85] "1979-02-16T00:00:00.000Z" "1979-05-16T00:00:00.000Z"
#>  [87] "1979-08-16T00:00:00.000Z" "1979-11-16T00:00:00.000Z"
#>  [89] "1980-02-16T00:00:00.000Z" "1980-05-16T00:00:00.000Z"
#>  [91] "1980-08-16T00:00:00.000Z" "1980-11-16T00:00:00.000Z"
#>  [93] "1981-02-16T00:00:00.000Z" "1981-05-16T00:00:00.000Z"
#>  [95] "1981-08-16T00:00:00.000Z" "1981-11-16T00:00:00.000Z"
#>  [97] "1982-02-16T00:00:00.000Z" "1982-05-16T00:00:00.000Z"
#>  [99] "1982-08-16T00:00:00.000Z" "1982-11-16T00:00:00.000Z"
#> [101] "1983-02-16T00:00:00.000Z" "1983-05-16T00:00:00.000Z"
#> [103] "1983-08-16T00:00:00.000Z" "1983-11-16T00:00:00.000Z"
#> [105] "1984-02-16T00:00:00.000Z" "1984-05-16T00:00:00.000Z"
#> [107] "1984-08-16T00:00:00.000Z" "1984-11-16T00:00:00.000Z"
#> [109] "1985-02-16T00:00:00.000Z" "1985-05-16T00:00:00.000Z"
#> [111] "1985-08-16T00:00:00.000Z" "1985-11-16T00:00:00.000Z"
#> [113] "1986-02-16T00:00:00.000Z" "1986-05-16T00:00:00.000Z"
#> [115] "1986-08-16T00:00:00.000Z" "1986-11-16T00:00:00.000Z"
#> [117] "1987-02-16T00:00:00.000Z" "1987-05-16T00:00:00.000Z"
#> [119] "1987-08-16T00:00:00.000Z" "1987-11-16T00:00:00.000Z"
#> [121] "1988-02-16T00:00:00.000Z" "1988-05-16T00:00:00.000Z"
#> [123] "1988-08-16T00:00:00.000Z" "1988-11-16T00:00:00.000Z"
#> [125] "1989-02-16T00:00:00.000Z" "1989-05-16T00:00:00.000Z"
#> [127] "1989-08-16T00:00:00.000Z" "1989-11-16T00:00:00.000Z"
#> [129] "1990-02-16T00:00:00.000Z" "1990-05-16T00:00:00.000Z"
#> [131] "1990-08-16T00:00:00.000Z" "1990-11-16T00:00:00.000Z"
#> [133] "1991-02-16T00:00:00.000Z" "1991-05-16T00:00:00.000Z"
#> [135] "1991-08-16T00:00:00.000Z" "1991-11-16T00:00:00.000Z"
#> [137] "1992-02-16T00:00:00.000Z" "1992-05-16T00:00:00.000Z"
#> [139] "1992-08-16T00:00:00.000Z" "1992-11-16T00:00:00.000Z"
#> [141] "1993-02-16T00:00:00.000Z" "1993-05-16T00:00:00.000Z"
#> [143] "1993-08-16T00:00:00.000Z" "1993-11-16T00:00:00.000Z"
#> [145] "1994-02-16T00:00:00.000Z" "1994-05-16T00:00:00.000Z"
#> [147] "1994-08-16T00:00:00.000Z" "1994-11-16T00:00:00.000Z"
#> [149] "1995-02-16T00:00:00.000Z" "1995-05-16T00:00:00.000Z"
#> [151] "1995-08-16T00:00:00.000Z" "1995-11-16T00:00:00.000Z"
#> [153] "1996-02-16T00:00:00.000Z" "1996-05-16T00:00:00.000Z"
#> [155] "1996-08-16T00:00:00.000Z" "1996-11-16T00:00:00.000Z"
#> [157] "1997-02-16T00:00:00.000Z" "1997-05-16T00:00:00.000Z"
#> [159] "1997-08-16T00:00:00.000Z" "1997-11-16T00:00:00.000Z"
#> [161] "1998-02-16T00:00:00.000Z" "1998-05-16T00:00:00.000Z"
#> [163] "1998-08-16T00:00:00.000Z" "1998-11-16T00:00:00.000Z"
#> [165] "1999-02-16T00:00:00.000Z" "1999-05-16T00:00:00.000Z"
#> [167] "1999-08-16T00:00:00.000Z" "1999-11-16T00:00:00.000Z"
#> [169] "2000-02-16T00:00:00.000Z" "2000-05-16T00:00:00.000Z"
#> [171] "2000-08-16T00:00:00.000Z" "2000-11-16T00:00:00.000Z"
#> [173] "2001-02-16T00:00:00.000Z" "2001-05-16T00:00:00.000Z"
#> [175] "2001-08-16T00:00:00.000Z" "2001-11-16T00:00:00.000Z"
#> [177] "2002-02-16T00:00:00.000Z" "2002-05-16T00:00:00.000Z"
#> [179] "2002-08-16T00:00:00.000Z" "2002-11-16T00:00:00.000Z"
#> [181] "2003-02-16T00:00:00.000Z" "2003-05-16T00:00:00.000Z"
#> [183] "2003-08-16T00:00:00.000Z" "2003-11-16T00:00:00.000Z"
#> [185] "2004-02-16T00:00:00.000Z" "2004-05-16T00:00:00.000Z"
#> [187] "2004-08-16T00:00:00.000Z" "2004-11-16T00:00:00.000Z"
#> [189] "2005-02-16T00:00:00.000Z" "2005-05-16T00:00:00.000Z"
#> [191] "2005-08-16T00:00:00.000Z" "2005-11-16T00:00:00.000Z"
#> [193] "2006-02-16T00:00:00.000Z" "2006-05-16T00:00:00.000Z"
#> [195] "2006-08-16T00:00:00.000Z" "2006-11-16T00:00:00.000Z"
#> [197] "2007-02-16T00:00:00.000Z" "2007-05-16T00:00:00.000Z"
#> [199] "2007-08-16T00:00:00.000Z" "2007-11-16T00:00:00.000Z"
#> [201] "2008-02-16T00:00:00.000Z" "2008-05-16T00:00:00.000Z"
#> [203] "2008-08-16T00:00:00.000Z" "2008-11-16T00:00:00.000Z"
#> [205] "2009-02-16T00:00:00.000Z" "2009-05-16T00:00:00.000Z"
#> [207] "2009-08-16T00:00:00.000Z" "2009-11-16T00:00:00.000Z"
#> [209] "2010-02-16T00:00:00.000Z" "2010-05-16T00:00:00.000Z"
#> [211] "2010-08-16T00:00:00.000Z" "2010-11-16T00:00:00.000Z"
#> [213] "2011-02-16T00:00:00.000Z" "2011-05-16T00:00:00.000Z"
#> [215] "2011-08-16T00:00:00.000Z" "2011-11-16T00:00:00.000Z"
#> [217] "2012-02-16T00:00:00.000Z" "2012-05-16T00:00:00.000Z"
#> [219] "2012-08-16T00:00:00.000Z" "2012-11-16T00:00:00.000Z"
#> [221] "2013-02-16T00:00:00.000Z" "2013-05-16T00:00:00.000Z"
#> [223] "2013-08-16T00:00:00.000Z" "2013-11-16T00:00:00.000Z"
#> [225] "2014-02-16T00:00:00.000Z" "2014-05-16T00:00:00.000Z"
#> [227] "2014-08-16T00:00:00.000Z" "2014-11-16T00:00:00.000Z"
#> [229] "2015-02-16T00:00:00.000Z" "2015-05-16T00:00:00.000Z"
#> [231] "2015-08-16T00:00:00.000Z" "2015-11-16T00:00:00.000Z"
#> [233] "2016-02-16T00:00:00.000Z" "2016-05-16T00:00:00.000Z"
#> [235] "2016-08-16T00:00:00.000Z" "2016-11-16T00:00:00.000Z"
#> attr(,"type")
#> temporal_coefficents
#> 
#> $Emodnetbio__aca_spp_19582016_L1
#>   [1] "1958-02-16T00:00:00.000Z" "1958-05-16T00:00:00.000Z"
#>   [3] "1958-08-16T00:00:00.000Z" "1958-11-16T00:00:00.000Z"
#>   [5] "1959-02-16T00:00:00.000Z" "1959-05-16T00:00:00.000Z"
#>   [7] "1959-08-16T00:00:00.000Z" "1959-11-16T00:00:00.000Z"
#>   [9] "1960-02-16T00:00:00.000Z" "1960-05-16T00:00:00.000Z"
#>  [11] "1960-08-16T00:00:00.000Z" "1960-11-16T00:00:00.000Z"
#>  [13] "1961-02-16T00:00:00.000Z" "1961-05-16T00:00:00.000Z"
#>  [15] "1961-08-16T00:00:00.000Z" "1961-11-16T00:00:00.000Z"
#>  [17] "1962-02-16T00:00:00.000Z" "1962-05-16T00:00:00.000Z"
#>  [19] "1962-08-16T00:00:00.000Z" "1962-11-16T00:00:00.000Z"
#>  [21] "1963-02-16T00:00:00.000Z" "1963-05-16T00:00:00.000Z"
#>  [23] "1963-08-16T00:00:00.000Z" "1963-11-16T00:00:00.000Z"
#>  [25] "1964-02-16T00:00:00.000Z" "1964-05-16T00:00:00.000Z"
#>  [27] "1964-08-16T00:00:00.000Z" "1964-11-16T00:00:00.000Z"
#>  [29] "1965-02-16T00:00:00.000Z" "1965-05-16T00:00:00.000Z"
#>  [31] "1965-08-16T00:00:00.000Z" "1965-11-16T00:00:00.000Z"
#>  [33] "1966-02-16T00:00:00.000Z" "1966-05-16T00:00:00.000Z"
#>  [35] "1966-08-16T00:00:00.000Z" "1966-11-16T00:00:00.000Z"
#>  [37] "1967-02-16T00:00:00.000Z" "1967-05-16T00:00:00.000Z"
#>  [39] "1967-08-16T00:00:00.000Z" "1967-11-16T00:00:00.000Z"
#>  [41] "1968-02-16T00:00:00.000Z" "1968-05-16T00:00:00.000Z"
#>  [43] "1968-08-16T00:00:00.000Z" "1968-11-16T00:00:00.000Z"
#>  [45] "1969-02-16T00:00:00.000Z" "1969-05-16T00:00:00.000Z"
#>  [47] "1969-08-16T00:00:00.000Z" "1969-11-16T00:00:00.000Z"
#>  [49] "1970-02-16T00:00:00.000Z" "1970-05-16T00:00:00.000Z"
#>  [51] "1970-08-16T00:00:00.000Z" "1970-11-16T00:00:00.000Z"
#>  [53] "1971-02-16T00:00:00.000Z" "1971-05-16T00:00:00.000Z"
#>  [55] "1971-08-16T00:00:00.000Z" "1971-11-16T00:00:00.000Z"
#>  [57] "1972-02-16T00:00:00.000Z" "1972-05-16T00:00:00.000Z"
#>  [59] "1972-08-16T00:00:00.000Z" "1972-11-16T00:00:00.000Z"
#>  [61] "1973-02-16T00:00:00.000Z" "1973-05-16T00:00:00.000Z"
#>  [63] "1973-08-16T00:00:00.000Z" "1973-11-16T00:00:00.000Z"
#>  [65] "1974-02-16T00:00:00.000Z" "1974-05-16T00:00:00.000Z"
#>  [67] "1974-08-16T00:00:00.000Z" "1974-11-16T00:00:00.000Z"
#>  [69] "1975-02-16T00:00:00.000Z" "1975-05-16T00:00:00.000Z"
#>  [71] "1975-08-16T00:00:00.000Z" "1975-11-16T00:00:00.000Z"
#>  [73] "1976-02-16T00:00:00.000Z" "1976-05-16T00:00:00.000Z"
#>  [75] "1976-08-16T00:00:00.000Z" "1976-11-16T00:00:00.000Z"
#>  [77] "1977-02-16T00:00:00.000Z" "1977-05-16T00:00:00.000Z"
#>  [79] "1977-08-16T00:00:00.000Z" "1977-11-16T00:00:00.000Z"
#>  [81] "1978-02-16T00:00:00.000Z" "1978-05-16T00:00:00.000Z"
#>  [83] "1978-08-16T00:00:00.000Z" "1978-11-16T00:00:00.000Z"
#>  [85] "1979-02-16T00:00:00.000Z" "1979-05-16T00:00:00.000Z"
#>  [87] "1979-08-16T00:00:00.000Z" "1979-11-16T00:00:00.000Z"
#>  [89] "1980-02-16T00:00:00.000Z" "1980-05-16T00:00:00.000Z"
#>  [91] "1980-08-16T00:00:00.000Z" "1980-11-16T00:00:00.000Z"
#>  [93] "1981-02-16T00:00:00.000Z" "1981-05-16T00:00:00.000Z"
#>  [95] "1981-08-16T00:00:00.000Z" "1981-11-16T00:00:00.000Z"
#>  [97] "1982-02-16T00:00:00.000Z" "1982-05-16T00:00:00.000Z"
#>  [99] "1982-08-16T00:00:00.000Z" "1982-11-16T00:00:00.000Z"
#> [101] "1983-02-16T00:00:00.000Z" "1983-05-16T00:00:00.000Z"
#> [103] "1983-08-16T00:00:00.000Z" "1983-11-16T00:00:00.000Z"
#> [105] "1984-02-16T00:00:00.000Z" "1984-05-16T00:00:00.000Z"
#> [107] "1984-08-16T00:00:00.000Z" "1984-11-16T00:00:00.000Z"
#> [109] "1985-02-16T00:00:00.000Z" "1985-05-16T00:00:00.000Z"
#> [111] "1985-08-16T00:00:00.000Z" "1985-11-16T00:00:00.000Z"
#> [113] "1986-02-16T00:00:00.000Z" "1986-05-16T00:00:00.000Z"
#> [115] "1986-08-16T00:00:00.000Z" "1986-11-16T00:00:00.000Z"
#> [117] "1987-02-16T00:00:00.000Z" "1987-05-16T00:00:00.000Z"
#> [119] "1987-08-16T00:00:00.000Z" "1987-11-16T00:00:00.000Z"
#> [121] "1988-02-16T00:00:00.000Z" "1988-05-16T00:00:00.000Z"
#> [123] "1988-08-16T00:00:00.000Z" "1988-11-16T00:00:00.000Z"
#> [125] "1989-02-16T00:00:00.000Z" "1989-05-16T00:00:00.000Z"
#> [127] "1989-08-16T00:00:00.000Z" "1989-11-16T00:00:00.000Z"
#> [129] "1990-02-16T00:00:00.000Z" "1990-05-16T00:00:00.000Z"
#> [131] "1990-08-16T00:00:00.000Z" "1990-11-16T00:00:00.000Z"
#> [133] "1991-02-16T00:00:00.000Z" "1991-05-16T00:00:00.000Z"
#> [135] "1991-08-16T00:00:00.000Z" "1991-11-16T00:00:00.000Z"
#> [137] "1992-02-16T00:00:00.000Z" "1992-05-16T00:00:00.000Z"
#> [139] "1992-08-16T00:00:00.000Z" "1992-11-16T00:00:00.000Z"
#> [141] "1993-02-16T00:00:00.000Z" "1993-05-16T00:00:00.000Z"
#> [143] "1993-08-16T00:00:00.000Z" "1993-11-16T00:00:00.000Z"
#> [145] "1994-02-16T00:00:00.000Z" "1994-05-16T00:00:00.000Z"
#> [147] "1994-08-16T00:00:00.000Z" "1994-11-16T00:00:00.000Z"
#> [149] "1995-02-16T00:00:00.000Z" "1995-05-16T00:00:00.000Z"
#> [151] "1995-08-16T00:00:00.000Z" "1995-11-16T00:00:00.000Z"
#> [153] "1996-02-16T00:00:00.000Z" "1996-05-16T00:00:00.000Z"
#> [155] "1996-08-16T00:00:00.000Z" "1996-11-16T00:00:00.000Z"
#> [157] "1997-02-16T00:00:00.000Z" "1997-05-16T00:00:00.000Z"
#> [159] "1997-08-16T00:00:00.000Z" "1997-11-16T00:00:00.000Z"
#> [161] "1998-02-16T00:00:00.000Z" "1998-05-16T00:00:00.000Z"
#> [163] "1998-08-16T00:00:00.000Z" "1998-11-16T00:00:00.000Z"
#> [165] "1999-02-16T00:00:00.000Z" "1999-05-16T00:00:00.000Z"
#> [167] "1999-08-16T00:00:00.000Z" "1999-11-16T00:00:00.000Z"
#> [169] "2000-02-16T00:00:00.000Z" "2000-05-16T00:00:00.000Z"
#> [171] "2000-08-16T00:00:00.000Z" "2000-11-16T00:00:00.000Z"
#> [173] "2001-02-16T00:00:00.000Z" "2001-05-16T00:00:00.000Z"
#> [175] "2001-08-16T00:00:00.000Z" "2001-11-16T00:00:00.000Z"
#> [177] "2002-02-16T00:00:00.000Z" "2002-05-16T00:00:00.000Z"
#> [179] "2002-08-16T00:00:00.000Z" "2002-11-16T00:00:00.000Z"
#> [181] "2003-02-16T00:00:00.000Z" "2003-05-16T00:00:00.000Z"
#> [183] "2003-08-16T00:00:00.000Z" "2003-11-16T00:00:00.000Z"
#> [185] "2004-02-16T00:00:00.000Z" "2004-05-16T00:00:00.000Z"
#> [187] "2004-08-16T00:00:00.000Z" "2004-11-16T00:00:00.000Z"
#> [189] "2005-02-16T00:00:00.000Z" "2005-05-16T00:00:00.000Z"
#> [191] "2005-08-16T00:00:00.000Z" "2005-11-16T00:00:00.000Z"
#> [193] "2006-02-16T00:00:00.000Z" "2006-05-16T00:00:00.000Z"
#> [195] "2006-08-16T00:00:00.000Z" "2006-11-16T00:00:00.000Z"
#> [197] "2007-02-16T00:00:00.000Z" "2007-05-16T00:00:00.000Z"
#> [199] "2007-08-16T00:00:00.000Z" "2007-11-16T00:00:00.000Z"
#> [201] "2008-02-16T00:00:00.000Z" "2008-05-16T00:00:00.000Z"
#> [203] "2008-08-16T00:00:00.000Z" "2008-11-16T00:00:00.000Z"
#> [205] "2009-02-16T00:00:00.000Z" "2009-05-16T00:00:00.000Z"
#> [207] "2009-08-16T00:00:00.000Z" "2009-11-16T00:00:00.000Z"
#> [209] "2010-02-16T00:00:00.000Z" "2010-05-16T00:00:00.000Z"
#> [211] "2010-08-16T00:00:00.000Z" "2010-11-16T00:00:00.000Z"
#> [213] "2011-02-16T00:00:00.000Z" "2011-05-16T00:00:00.000Z"
#> [215] "2011-08-16T00:00:00.000Z" "2011-11-16T00:00:00.000Z"
#> [217] "2012-02-16T00:00:00.000Z" "2012-05-16T00:00:00.000Z"
#> [219] "2012-08-16T00:00:00.000Z" "2012-11-16T00:00:00.000Z"
#> [221] "2013-02-16T00:00:00.000Z" "2013-05-16T00:00:00.000Z"
#> [223] "2013-08-16T00:00:00.000Z" "2013-11-16T00:00:00.000Z"
#> [225] "2014-02-16T00:00:00.000Z" "2014-05-16T00:00:00.000Z"
#> [227] "2014-08-16T00:00:00.000Z" "2014-11-16T00:00:00.000Z"
#> [229] "2015-02-16T00:00:00.000Z" "2015-05-16T00:00:00.000Z"
#> [231] "2015-08-16T00:00:00.000Z" "2015-11-16T00:00:00.000Z"
#> [233] "2016-02-16T00:00:00.000Z" "2016-05-16T00:00:00.000Z"
#> [235] "2016-08-16T00:00:00.000Z" "2016-11-16T00:00:00.000Z"
#> attr(,"type")
#> temporal_coefficents
#>